Split h and patch into separate files.
- Fix prop diffing for value and checked attributes. Close #6.
